refactor: extract anchor safety owners
Problem: the implementation added lines to several already-baselined oversized files and failed the repository lint gate.\n\nRoot cause: new anchor audit, gate, config rendering, ledger, and batch tests were placed in legacy owner files.\n\nFix: move each concern into a focused sibling file and compact the user-global config preservation without changing runtime behavior or updating lint baselines.\n\nVerification: go test ./...; go test -race ./internal/agent ./internal/evidence ./cmd/e2ebench; go vet ./...; go run ./tools/repolint.
