+++ internal/agent/evidence_flow_test.go
@@ -1140,125 +1140,6 @@ func TestEvidenceFlowAllowsBatchCompleteStepSignoffs(t *testing.T) {
 	}
 }
 
-func TestEvidenceFlowAllowsThreeOrderedCompleteStepSignoffs(t *testing.T) {
-	todoWrite, ok := tool.LookupBuiltin("todo_write")
-	if !ok {
-		t.Fatal("todo_write builtin not registered")
-	}
-	completeStep, ok := tool.LookupBuiltin("complete_step")
-	if !ok {
-		t.Fatal("complete_step builtin not registered")
-	}
-	reg := tool.NewRegistry()
-	reg.Add(todoWrite)
-	reg.Add(completeStep)
-	prov := &scriptedProvider{name: "p", turns: [][]provider.Chunk{
-		{
-			toolCallChunk("t", "todo_write", `{"todos":[
-				{"content":"one","status":"in_progress"},
-				{"content":"two","status":"pending"},
-				{"content":"three","status":"pending"}
-			]}`),
-			{Type: provider.ChunkDone},
-		},
-		{
-			toolCallChunk("c1", "complete_step", `{"step":"one","result":"done","evidence":[{"kind":"manual","summary":"checked"}]}`),
-			toolCallChunk("c2", "complete_step", `{"step":"two","result":"done","evidence":[{"kind":"manual","summary":"checked"}]}`),
-			toolCallChunk("c3", "complete_step", `{"step":"three","result":"done","evidence":[{"kind":"manual","summary":"checked"}]}`),
-			{Type: provider.ChunkDone},
-		},
-		{{Type: provider.ChunkText, Text: "done"}, {Type: provider.ChunkDone}},
-	}}
-	a := New(prov, reg, NewSession(""), Options{}, event.Discard)
-	if err := a.Run(withNoClosedLoop(context.Background()), "sign off three ordered steps"); err != nil {
-		t.Fatalf("Run: %v", err)
-	}
-	if got := len(toolResults(a.sess.conversation, "complete_step")); got != 3 {
-		t.Fatalf("complete_step results = %d, want three in one batch", got)
-	}
-	if prov.call != 3 {
-		t.Fatalf("provider calls = %d, want setup, one batch, final answer", prov.call)
-	}
-	for i, todo := range a.CanonicalTodoState() {
-		if todo.Status != "completed" {
-			t.Fatalf("canonical todo %d = %+v, want completed", i+1, todo)
-		}
-	}
-}
-
-func TestEvidenceFlowBatchCompletesPhaseChildrenThenPhaseHeader(t *testing.T) {
-	todoWrite, ok := tool.LookupBuiltin("todo_write")
-	if !ok {
-		t.Fatal("todo_write builtin not registered")
-	}
-	completeStep, ok := tool.LookupBuiltin("complete_step")
-	if !ok {
-		t.Fatal("complete_step builtin not registered")
-	}
-	reg := tool.NewRegistry()
-	reg.Add(todoWrite)
-	reg.Add(completeStep)
-	manual := `[{"kind":"manual","summary":"checked"}]`
-	prov := &scriptedProvider{name: "p", turns: [][]provider.Chunk{
-		{toolCallChunk("t", "todo_write", `{"todos":[
-			{"content":"Phase","status":"pending","level":0},
-			{"content":"child one","status":"in_progress","level":1},
-			{"content":"child two","status":"pending","level":1},
-			{"content":"Finalize","status":"pending","level":0}
-		]}`), {Type: provider.ChunkDone}},
-		{
-			toolCallChunk("c1", "complete_step", `{"step":"child one","result":"done","evidence":`+manual+`}`),
-			toolCallChunk("c2", "complete_step", `{"step":"child two","result":"done","evidence":`+manual+`}`),
-			toolCallChunk("c3", "complete_step", `{"step":"Phase","result":"done","evidence":`+manual+`}`),
-			toolCallChunk("c4", "complete_step", `{"step":"Finalize","result":"done","evidence":`+manual+`}`),
-			{Type: provider.ChunkDone},
-		},
-		{{Type: provider.ChunkText, Text: "done"}, {Type: provider.ChunkDone}},
-	}}
-	a := New(prov, reg, NewSession(""), Options{}, event.Discard)
-	if err := a.Run(withNoClosedLoop(context.Background()), "sign off phase children and header"); err != nil {
-		t.Fatalf("Run: %v", err)
-	}
-	if got := len(toolResults(a.sess.conversation, "complete_step")); got != 4 {
-		t.Fatalf("complete_step results = %d, want four ordered sign-offs", got)
-	}
-	for i, todo := range a.CanonicalTodoState() {
-		if todo.Status != "completed" {
-			t.Fatalf("canonical todo %d = %+v, want completed", i+1, todo)
-		}
-	}
-}
-
-func TestEvidenceFlowMixedBatchKeepsReceiptOrderAcrossSignoffs(t *testing.T) {
-	reg := evidenceRegistry()
-	prov := &scriptedProvider{name: "p", turns: [][]provider.Chunk{
-		{toolCallChunk("t", "todo_write", `{"todos":[
-			{"content":"one","status":"in_progress"},
-			{"content":"two","status":"pending"}
-		]}`), {Type: provider.ChunkDone}},
-		{
-			toolCallChunk("b1", "bash", `{"command":"go test ./one"}`),
-			toolCallChunk("c1", "complete_step", `{"step":"one","result":"verified","evidence":[{"kind":"verification","summary":"one passed","command":"go test ./one"}]}`),
-			toolCallChunk("b2", "bash", `{"command":"go test ./two"}`),
-			toolCallChunk("c2", "complete_step", `{"step":"two","result":"verified","evidence":[{"kind":"verification","summary":"two passed","command":"go test ./two"}]}`),
-			{Type: provider.ChunkDone},
-		},
-		{{Type: provider.ChunkText, Text: "done"}, {Type: provider.ChunkDone}},
-	}}
-	a := New(prov, reg, NewSession(""), Options{}, event.Discard)
-	if err := a.Run(withNoClosedLoop(context.Background()), "verify and sign off mixed work"); err != nil {
-		t.Fatalf("Run: %v", err)
-	}
-	if got := len(toolResults(a.sess.conversation, "complete_step")); got != 2 {
-		t.Fatalf("complete_step results = %d, want two", got)
-	}
-	for i, todo := range a.CanonicalTodoState() {
-		if todo.Status != "completed" {
-			t.Fatalf("canonical todo %d = %+v, want completed", i+1, todo)
-		}
-	}
-}
-
 func TestEvidenceFlowFailedCompleteStepDoesNotAuthorizeTodoCompletion(t *testing.T) {
 	todoWrite, ok := tool.LookupBuiltin("todo_write")
 	if !ok {
