+++ internal/config/config.go
@@ -1327,10 +1327,7 @@ type AgentConfig struct {
 	// PlanModeReadOnlyCommands is retained for old config/session round trips. Main
 	// Plan bash calls now use the ordinary Permissions classifier and Sandbox.
 	PlanModeReadOnlyCommands []string `toml:"plan_mode_read_only_commands"`
-	// LegacyAnchorSafetyGate keeps delete_range on the pre-fingerprint full-read
-	// guard. It is an operational rollback switch; the default is false so the
-	// interval-fingerprint gate is active.
-	LegacyAnchorSafetyGate bool `toml:"legacy_anchor_safety_gate"`
+	LegacyAnchorSafetyGate   bool     `toml:"legacy_anchor_safety_gate"` // user-global rollback to the full-read guard
 }
 
 // ProviderEntry declares a model provider instance. ContextWindow is the model's
