+++ internal/config/load.go
@@ -133,8 +133,7 @@ func loadForRoot(root string, migrateOnDisk bool) (*Config, error) {
 	globalDesktopLanguage := cfg.Desktop.Language
 	globalPricingCurrency := cfg.Desktop.Currency
 	globalBillingDisplayCurrency := cfg.Billing.DisplayCurrency
-	globalTelemetry := cfg.Telemetry
-	globalLegacyAnchorSafetyGate := cfg.Agent.LegacyAnchorSafetyGate
+	globalTelemetry, globalLegacyAnchorSafetyGate := cfg.Telemetry, cfg.Agent.LegacyAnchorSafetyGate
 
 	tomlSources = append(tomlSources, projectTOML)
 	projectMeta, err := mergeTOML(cfg, projectTOML)
@@ -169,10 +168,7 @@ func loadForRoot(root string, migrateOnDisk bool) (*Config, error) {
 	cfg.Billing.DisplayCurrency = globalBillingDisplayCurrency
 	// CLI telemetry is an explicit user-global privacy choice. Project config
 	// cannot opt a user in or out, including when the global value is absent.
-	cfg.Telemetry = globalTelemetry
-	// The delete_range rollback switch is a user-global operational control;
-	// repository-local config must not weaken the workspace safety policy.
-	cfg.Agent.LegacyAnchorSafetyGate = globalLegacyAnchorSafetyGate
+	cfg.Telemetry, cfg.Agent.LegacyAnchorSafetyGate = globalTelemetry, globalLegacyAnchorSafetyGate
 	// TOML decoding replaces [[plugins]] wholesale, so cfg.Plugins now holds
 	// only the last file's. Re-merge by name across all sources (later wins) so a
 	// project reasonix.toml doesn't drop the global config's MCP servers.
